+++ paifeaturestore-20230621/src/Models/ListFeatureEntitiesRequest.php
@@ -49,8 +49,6 @@ class ListFeatureEntitiesRequest extends Model
     public $pageSize;
 
     /**
-     * @description This parameter is required.
-     *
      * @example 3
      *
      * @var string