+++ paifeaturestore-20230621/src/Models/ListFeatureEntitiesShrinkRequest.php
@@ -49,8 +49,6 @@ class ListFeatureEntitiesShrinkRequest extends Model
     public $pageSize;
 
     /**
-     * @description This parameter is required.
-     *
      * @example 3
      *
      * @var string