+++ paifeaturestore-20230621/src/Models/ListFeatureViewsRequest.php
@@ -56,8 +56,6 @@ class ListFeatureViewsRequest extends Model
     public $pageSize;
 
     /**
-     * @description This parameter is required.
-     *
      * @example 3
      *
      * @var string