+++ paifeaturestore-20230621/src/Models/ListFeatureViewsShrinkRequest.php
@@ -56,8 +56,6 @@ class ListFeatureViewsShrinkRequest extends Model
     public $pageSize;
 
     /**
-     * @description This parameter is required.
-     *
      * @example 3
      *
      * @var string