+++ paifeaturestore-20230621/src/Models/ListModelFeaturesRequest.php
@@ -49,8 +49,6 @@ class ListModelFeaturesRequest extends Model
     public $pageSize;
 
     /**
-     * @description This parameter is required.
-     *
      * @example 4
      *
      * @var string