+++ paifeaturestore-20230621/src/Models/ListTasksShrinkRequest.php
@@ -37,8 +37,6 @@ class ListTasksShrinkRequest extends Model
     public $pageSize;
 
     /**
-     * @description This parameter is required.
-     *
      * @example 4
      *
      * @var string