+++ paifeaturestore-20230621/ChangeLog.md
@@ -1,3 +1,15 @@
+2024-12-18 Version: 3.0.0
+- Delete API CreateModelFeatureTrainingSetFGTable.
+- Delete API ExportModelFeatureTrainingSetFGTable.
+- Delete API UpdateModelFeatureFGInfo.
+- Update API ListFeatureEntities: update param ProjectId.
+- Update API ListFeatureViews: update param ProjectId.
+- Update API ListLabelTables: update param ProjectId.
+- Update API ListModelFeatures: update param ProjectId.
+- Update API ListTasks: update param ProjectId.
+- Update API UpdateModelFeature: update param body.
+
+
 2024-10-16 Version: 2.0.1
 - Update API GetInstance: update response param.
 - Update API ListInstances: update response param.